Move the download Read impl into its own module
