+++ zcash_proofs/src/lib.rs
@@ -33,6 +33,10 @@ pub mod sprout;
 )]
 pub mod prover;
 
+#[cfg(feature = "download-params")]
+#[cfg_attr(docsrs, doc(cfg(feature = "download-params")))]
+mod downloadreader;
+
 // Circuit names
 
 /// The sapling spend parameters file name.
@@ -249,8 +253,8 @@ fn stream_params_downloads_to_disk(
 
     // Download the responses and write them to a new file,
     // verifying the hash as bytes are read.
-    let params_download_1 = ResponseLazyReader::from(params_download_1);
-    let params_download_2 = ResponseLazyReader::from(params_download_2);
+    let params_download_1 = downloadreader::ResponseLazyReader::from(params_download_1);
+    let params_download_2 = downloadreader::ResponseLazyReader::from(params_download_2);
 
     // Limit the download size to avoid DoS.
     // This also avoids launching the second request, if the first request provides enough bytes.
@@ -270,86 +274,6 @@ fn stream_params_downloads_to_disk(
     )?;
 
     Ok(())
-}
-
-/// A wrapper that implements [`io::Read`] on a [`minreq::ResponseLazy`].
-#[cfg(feature = "download-params")]
-enum ResponseLazyReader {
-    Request(minreq::Request),
-    Response(minreq::ResponseLazy),
-    Complete(Result<(), String>),
-}
-
-#[cfg(feature = "download-params")]
-impl From<minreq::Request> for ResponseLazyReader {
-    fn from(request: minreq::Request) -> ResponseLazyReader {
-        ResponseLazyReader::Request(request)
-    }
-}
-
-#[cfg(feature = "download-params")]
-impl From<minreq::ResponseLazy> for ResponseLazyReader {
-    fn from(response: minreq::ResponseLazy) -> ResponseLazyReader {
-        ResponseLazyReader::Response(response)
-    }
-}
-
-#[cfg(feature = "download-params")]
-impl io::Read for ResponseLazyReader {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        use ResponseLazyReader::*;
-
-        // Zero-sized buffer. This should never happen.
-        if buf.is_empty() {
-            return Ok(0);
-        }
-
-        loop {
-            match self {
-                // Launch a lazy response for this request
-                Request(request) => match request.clone().send_lazy() {
-                    Ok(response) => *self = Response(response),
-                    Err(error) => {
-                        let error = Err(format!("download request failed: {:?}", error));
-
-                        *self = Complete(error);
-                    }
-                },
-
-                // Read from the response
-                Response(response) => {
-                    // minreq has a very limited lazy reading interface.
-                    match &mut response.next() {
-                        // Read one byte into the buffer.
-                        // We ignore the expected length, because we have no way of telling the BufReader.
-                        Some(Ok((byte, _length))) => {
-                            buf[0] = *byte;
-                            return Ok(1);
-                        }
-
-                        // Reading failed.
-                        Some(Err(error)) => {
-                            let error = Err(format!("download response failed: {:?}", error));
-
-                            *self = Complete(error);
-                        }
-
-                        // Finished reading.
-                        None => *self = Complete(Ok(())),
-                    }
-                }
-
-                Complete(result) => {
-                    return match result {
-                        // Return a zero-byte read for download success and EOF.
-                        Ok(()) => Ok(0),
-                        // Keep returning the download error,
-                        Err(error) => Err(io::Error::new(io::ErrorKind::Other, error.clone())),
-                    };
-                }
-            }
-        }
-    }
 }
 
 /// Zcash Sprout and Sapling groth16 circuit parameters.
